Added ErrorHandling, notFound middlewares and mergeImages controller.

+++ src/controllers/deleteById.js
@@ -1,7 +1,19 @@
 const database = require("../entities/Database")
+const { NotFoundApiError } = require("../validators/ApiError")
 
-module.exports = async (req, res) => {
-  const imageId = req.params.id
-  const id = await database.remove(imageId)
-  return res.json({ id })
+module.exports = async (req, res, next) => {
+  try {
+    const imageId = req.params.id
+
+    const image = database.findById(imageId)
+
+    if (!image) {
+      throw new NotFoundApiError("Image associated with this ID was not found")
+    }
+
+    const id = await database.remove(imageId)
+    return res.json({ id })
+  } catch (err) {
+    return next(err)
+  }
 }
